code optimized added middleware bugs fixed inputs checked new schedule in lessons.txt auto users adding

+++ bot/handlers.py
@@ -3,12 +3,11 @@ import time
 import telebot.apihelper
 from telebot import types
 
-from config import cabinets_info, bot, ADMIN_ID, pages, days, define_time
-from sql_requests import get_teacher, create_lesson, update_user_level
+from config import cabinets_info, bot, ADMIN_ID, pages, days, define_time, GROUP_ID
+from sql_requests import get_teacher, create_lesson, update_user_level, delete_lesson
 from utils import send_weather, detect_user, check_permissions, random_element, format_schedule, format_teacher
 
 
-# ready to use
 def start_greetings(message: types.Message):
     bot.send_message(ADMIN_ID, f"{detect_user(message)} запустил бота")
 
@@ -24,30 +23,36 @@ def start_greetings(message: types.Message):
     bot.send_message(message.chat.id, pages["home"][1], parse_mode="html", reply_markup=markup)
 
 
+# ready
 def set_permission(message: types.Message):
     if check_permissions(message, 5):
 
         request = message.text.split()[1:3]
 
         try:
-            request[1] = min(int(request[1]), 5)
+            request[1] = max(min(int(request[1]), 5), -1)
+
+            request[0] = request[0].replace('@', '')
+
         except (IndexError, ValueError):
             bot.send_message(message.chat.id,
                              "<b>❌ Неверная команда!</b>\n"
-                             "<i>Используйте</i> <code>/set |юзернейм| |уровень|</code>\n"
+                             "<i>Используйте</i> <code>/set |имя пользователя| |уровень|</code>\n"
                              "Подробнее: /help", parse_mode="html")
             return
+        bot.send_message(message.chat.id,
+                         f"<b>✅ Уровень пользователя <i>{'' if request[0].isdigit() else '@'}{request[0]}"
+                         f"</i> теперь {request[1]}</b>",
+                         parse_mode='html')
 
         update_user_level(request[0], request[1])
-    else:
-        bot.set_message_reaction(message.chat.id, message.message_id, [types.ReactionTypeEmoji("😨")])
 
 
-# ready to use
+# ready
 def manage_cabs(message: types.Message):
     args = message.text.split()[1:5]
     print(f"[=]{detect_user(message)} requested cabinets: {args}")
-    args = [i[:4] for i in args if len(i) >= 3 and i[:3].isdigit()]
+    args = [i[:4] for i in args if len(i) >= 3 and i[:3].isdigit() and not i[3:].isdigit()]
 
     if len(args) > 0:
         cabinets_info['cabinets'] = args
@@ -63,10 +68,10 @@ def manage_cabs(message: types.Message):
     else:
         bot.send_message(message.chat.id, "<b>Кабинеты не добавлены</b>", parse_mode="html")
 
-    bot.delete_message(message.chat.id, message.message_id)
+    # bot.delete_message(message.chat.id, message.message_id)
 
 
-# ready to use
+# ready
 def show_author(callback: types.CallbackQuery):
     bot.answer_callback_query(callback.id,
                               f"Прислал: {cabinets_info['author'].get(callback.message.message_id - 1, 'неизвестно')}",
@@ -74,16 +79,19 @@ def show_author(callback: types.CallbackQuery):
 
 
 # ready to use
-def send_schedule(message: types.Message):
-    args = message.text.split()[1:3]
+def send_schedule(message: types.Message = None):
+    if not message:
+        bot.send_message(GROUP_ID, format_schedule(), parse_mode='html')
+        return
 
+    args = message.text.split()[1:3]
     print(f"[=]{detect_user(message)} requested schedule: {args}")
 
     args.extend([''] * (2 - len(args)))
     week = list(filter(lambda x: x.isdigit(), args))
-    week = int(week[0]) % 20 if week else None
+    week = int(week[0]) % 21 if week else None
     day_of_the_week = list(filter(lambda x: x in days, args))
-    day_of_the_week = days.index(day_of_the_week[0]) if day_of_the_week else None
+    day_of_the_week = days.get(day_of_the_week[0]) if day_of_the_week else None
 
     markup = types.InlineKeyboardMarkup()
     if day_of_the_week is None:
@@ -93,17 +101,29 @@ def send_schedule(message: types.Message):
 
     bot.send_message(message.chat.id, format_schedule(week, day_of_the_week), parse_mode='html',
                      reply_markup=markup)
+    bot.delete_message(message.chat.id, message.message_id)
 
 
 def add_lesson(message: types.Message):
     if check_permissions(message, 4):
         args = message.text.split()[1:]
-
         try:
-            args[1] = days.index(args[1])
-            args[2] = int(args[2])
-            args[3] = args[3].capitalize().replace('_', ' ')
-            args[5] = args[5].capitalize()
+
+            interval = args[0].split('-')
+            start = int(interval[0])
+            end = int(interval[0] if len(interval) <= 1 else interval[1])
+
+            if 0 <= start <= 20 and start <= end <= 20:
+
+                args[1] = days.get(args[1])
+                args[2] = int(args[2]) - 1
+                args[3] = args[3].capitalize().replace('_', ' ')
+                if len(args) > 5:
+                    args[5] = args[5].capitalize()
+            else:
+                bot.set_message_reaction(message.chat.id, message.message_id, [types.ReactionTypeEmoji("😨")])
+                return
+
         except (IndexError, ValueError):
             print(f"[!]{detect_user(message)} WRONG ARGUMENTS: {args}")
 
@@ -116,28 +136,61 @@ def add_lesson(message: types.Message):
             create_lesson(*args[:6])
             bot.send_message(message.chat.id,
                              f"ℹ️ <b>{detect_user(message)}</b> добавил(-а):\n"
-                             f" - <b><em>{args[4] if args[4] == '-' else ''} {args[3]}</em></b>"
-                             f" в <b>{define_time[args[1]]}</b> на <b>{args[0]}</b> недели(-ю)",
+                             f" - <b><em> {args[4] if len(args) > 4 else 'доп.'} {args[3]}</em></b>"
+                             f" в <b>{days.get(args[1])} {define_time[args[2]]}</b> на <b>{args[0]}</b> неделю(-и)",
                              parse_mode="html")
 
+        ...  # wrong weeks
+
+        # bot.delete_message(message.chat.id, message.message_id)
+
+
+def remove_lesson(message: types.Message):
+    if check_permissions(message, 4):
+        args = message.text.split()[1:]
+        try:
+            args[0] = days.get(args[0])
+            args[1] = int(args[1]) - 1
+            args[2] = args[2].replace('_', ' ')
+
+        except (IndexError, ValueError):
+            print(f"[!]{detect_user(message)} WRONG ARGUMENTS: {args}")
+
+            bot.send_message(message.chat.id,
+                             "<b>❌ Неверная команда!</b>\n"
+                             "<i>Используйте</i> <code>/remove пн 1 Название_пары </code>"
+                             "Подробнее: /help", parse_mode="html")
+        else:
+            data = delete_lesson(*args[:3])
+            print(data)
+            if len(data) == 1:
+                bot.send_message(message.chat.id,
+                                 f"ℹ️ <b>{detect_user(message)}</b> удалил(-а):\n"
+                                 f" - <b><em>{data[0][2]} {data[0][3]}</em></b>"
+                                 f" в <b>{define_time[data[0][1]]}</b> на "
+                                 f"<b>{data[0][4]}{'' if data[0][4] == data[0][5] else data[0][5]}</b> неделe(-ях)",
+                                 parse_mode="html")
+            else:
+                bot.send_message(message.chat.id,
+                                 f"<b>❌ {"Слишком много результатов!" if len(data) > 1 else "Нет результатов"}</b>\n",
+                                 parse_mode='html')
+
         bot.delete_message(message.chat.id, message.message_id)
-    else:
-        bot.set_message_reaction(message.chat.id, message.message_id, [types.ReactionTypeEmoji("😨")])
 
 
-# ready to use !!need formatting!!
 def send_teacher(message: types.Message):
-    args = message.text.split()[1:]
-    if args:
-        result = (f"Занятия по запросу '{args[0]}' :\n"
+    args = message.text.lower().split()[1:]
+
+    if args and len(args[0]) >= 3:
+        result = (f"🔍 Занятия по запросу '<u>{args[0]}</u>' :\n"
                   f"{format_teacher(get_teacher(args[0]))}")
-        bot.send_message(message.chat.id, result)
+        bot.send_message(message.chat.id, result, parse_mode="html")
+    else:
+        bot.set_message_reaction(message.chat.id, message.message_id, [types.ReactionTypeEmoji("👀")])
 
 
-# old
 def weather_request(message: types.Message):
-    print(
-        f"[=]{detect_user(message)} REQUESTED WEATHER")
+    print(f"[=]{detect_user(message)} REQUESTED WEATHER")
     send_weather(message.chat.id)
     bot.delete_message(message.chat.id, message.message_id)
 
@@ -146,16 +199,15 @@ def weather_request(message: types.Message):
 def random_request(message: types.Message):
     if check_permissions(message, 4):
         args = message.text.split()[1:]
-        if args:
+        if args and len(args) >= 2:
             bot.send_message(message.chat.id, f"Среди {', '.join(args)}, был случайно выбран {random_element(args)}")
         bot.delete_message(message.chat.id, message.message_id)
-    else:
-        bot.set_message_reaction(message.chat.id, message.message_id, [types.ReactionTypeEmoji("😨")])
 
 
 # old
-def send_user_id(message: types.Message):
-    bot.reply_to(message, message.from_user.id)
+def send_id(message: types.Message):
+    bot.reply_to(message,
+                 f"ID: {message.from_user.id}\n{f'CHAT: {message.chat.id}' if message.chat.type != 'private' else ''}")
 
 
 def wrong_chat_type(message: types.Message):
@@ -197,12 +249,14 @@ def send_guide(message: types.Message):
 
 def scroll_schedule(callback: types.CallbackQuery):
     week = int(callback.message.text.split()[4])
-    week = week % 20 + 1 if callback.data == "next" else week % 20 - 1
+    week = week - (-1 if callback.data == 'next' else 1)
+    week = (week - 1 if week == 0 else week) % 21
 
     result = format_schedule(week)
 
-    try:
-        bot.edit_message_text(result, callback.message.chat.id, callback.message.message_id,
-                              parse_mode="html", reply_markup=callback.message.reply_markup)
-    except telebot.apihelper.ApiTelegramException:
-        print(f"[!]{detect_user(callback)} TO MANY CALLBACK REQUESTS")
+    if result != callback.message.text:
+        try:
+            bot.edit_message_text(result, callback.message.chat.id, callback.message.message_id,
+                                  parse_mode="html", reply_markup=callback.message.reply_markup)
+        except telebot.apihelper.ApiTelegramException:
+            print(f"[!]{detect_user(callback)} TO MANY CALLBACK REQUESTS")
